🆕 feat(Slider): support for other numeric types including double (#1050)
